Some GRIB stuff

+++ Semestr IV/Programowanie Zaawasowane I/new_grib_decoder/src/convert.h
@@ -13,7 +13,7 @@
 
 namespace convert {
     uint32_t to_size(std::vector<uint8_t>::iterator index);
-    uint8_t to_byte(std::vector<uint8_t>::iterator index);
+    uint16_t to_byte(std::vector<uint8_t>::iterator index);
     int32_t to_long(std::vector<uint8_t>::iterator index);
     int16_t to_short(std::vector<uint8_t>::iterator index);
     float to_float(std::vector<uint8_t>::iterator index);
